Fixed building generate_persistent_state
