+++ test/generate_persistent_state/main.cpp
@@ -53,21 +53,6 @@ void WritePersistentStateToFile()
     fclose(f);
 }
 
-OSResult System::TakeSemaphore(OSSemaphoreHandle, std::chrono::milliseconds)
-{
-    return OSResult::Success;
-}
-
-OSResult System::GiveSemaphore(OSSemaphoreHandle)
-{
-    return OSResult::Success;
-}
-
-OSSemaphoreHandle System::CreateBinarySemaphore(std::uint8_t)
-{
-    return reinterpret_cast<OSSemaphoreHandle>(1);
-}
-
 int main()
 {
     __libc_init_array();
